fix(audio): prefer media framework playback, TTS lane first
audio_playback_open tried the direct NuttX audio device before
media_player — the exact inverse of the capture side, and of what the
media-first audio architecture wants. Reorder: media_player first, then
CONFIG_AI_AGENT_AUDIO_NUTTX_DIRECT as the fallback for boards with no
media server (BK7258 pcm0p). Log which backend won.

Within media_player, open the dedicated TTS lane (graph.conf
abufsrc@TTS -> alsasink@pcm0p/SCOtx) before the generic Music lane, so
a concurrent music/recorder session on Music cannot force-close our
player; Music remains the fallback for graphs without a TTS entry.

Handle partial writes: media_player_write_data may accept less than
asked (its socket loop stops at the first short send); re-feed the
remainder so a clip is not truncated mid-sentence.

Fix the weak media_player_write_data stub signature (int -> ssize_t) to
match media_player.h, and reword the AI_AGENT_AUDIO_NUTTX_DIRECT help:
it is a fallback for media-less boards, leave it off on media boards.

+++ src/stubs.c
@@ -273,7 +273,10 @@ __attribute__((weak)) bool agent_shutdown_requested(void)
 
 /* ── Media player write_data stub ─────────────────────────── */
 
-__attribute__((weak)) int media_player_write_data(void* player,
+/* Signature must match frameworks/multimedia/media/include/media_player.h
+ * (ssize_t) — an int stub silently truncates and breaks the partial-write
+ * retry loop in audio_playback.c. */
+__attribute__((weak)) ssize_t media_player_write_data(void* player,
     const void* data, size_t len)
 {
     (void)player;
